Login registration lecture complete

+++ week4/day1And2/mvcdemo/src/main/java/com/coreym/mvcdemo/controllers/MainController.java
@@ -15,8 +15,11 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import com.coreym.mvcdemo.models.Pizza;
+import com.coreym.mvcdemo.models.User;
 import com.coreym.mvcdemo.services.PizzaService;
+import com.coreym.mvcdemo.services.UserService;
 
+import jakarta.servlet.http.HttpSession;
 import jakarta.validation.Valid;
 
 @Controller
@@ -25,18 +28,36 @@ public class MainController {
 	@Autowired
 	public PizzaService pizzaService;
 	
+	@Autowired
+	public UserService userService;
+	
 	@GetMapping("/")
-	public String home(@ModelAttribute Pizza pizza, Model model) {
+	public String home(@ModelAttribute Pizza pizza, Model model, HttpSession session, RedirectAttributes redirectAttributes) {
+		
+		if (session.getAttribute("loggedInUser") == null) {
+			redirectAttributes.addFlashAttribute("notLoggedIn", "You must be logged in to view this page");
+			return "redirect:/auth/login";
+		}
 		// @ModelAttribute - creates an empty Pizza object to attach to our form
 		ArrayList<Pizza> allPizzas = pizzaService.all();
+		
+		// We need to type cast the id because everything in session is stored as a generic Object
+		User userFromDb = userService.findOne((Long) session.getAttribute("loggedInUser"));
+		
 		model.addAttribute("allPizzas",allPizzas);
+		model.addAttribute("loggedInUser", userFromDb.getFirstName());
+		
 		String[] pizzaSizes = {"Small", "Medium", "Large", "Extra Large"};
 		model.addAttribute("pizzaSizes", pizzaSizes);
 		return "index.jsp";
 	}
 	
 	@GetMapping("/pizzas/{id}")
-	public String displayPizza (Model model, @PathVariable Long id, RedirectAttributes redirectAttributes) {
+	public String displayPizza (Model model, @PathVariable Long id, RedirectAttributes redirectAttributes, HttpSession session) {
+		if (session.getAttribute("loggedInUser") == null) {
+			redirectAttributes.addFlashAttribute("notLoggedIn", "You must be logged in to view this page");
+			return "redirect:/auth/login";
+		}
 		Pizza onePizza = pizzaService.findOne(id);
 		
 		if (onePizza == null) {
@@ -52,8 +73,11 @@ public class MainController {
 	}
 	
 	@PostMapping("/pizzas")
-	public String processCreatePizza(@Valid @ModelAttribute Pizza pizza, BindingResult result, Model model) {
-		
+	public String processCreatePizza(@Valid @ModelAttribute Pizza pizza, BindingResult result, Model model, RedirectAttributes redirectAttributes, HttpSession session) {
+		if (session.getAttribute("loggedInUser") == null) {
+			redirectAttributes.addFlashAttribute("notLoggedIn", "You must be logged in to view this page");
+			return "redirect:/auth/login";
+		}
 		// @Valid -> Attaches to the model attribute and makes sure the user input matches validations
 		// @ModelAttribute -> uses the binded data object within this request
 		// BindingResult -> holds the form data 
@@ -73,7 +97,11 @@ public class MainController {
 	}
 	
 	@GetMapping("/pizzas/edit/{id}")
-	public String editPizzaForm(@PathVariable Long id, RedirectAttributes redirectAttributes, Model model) {
+	public String editPizzaForm(@PathVariable Long id, RedirectAttributes redirectAttributes, Model model, HttpSession session) {
+		if (session.getAttribute("loggedInUser") == null) {
+			redirectAttributes.addFlashAttribute("notLoggedIn", "You must be logged in to view this page");
+			return "redirect:/auth/login";
+		}
 		Pizza pizzaToUpdate = pizzaService.findOne(id);
 		
 		if (pizzaToUpdate == null) {
@@ -91,8 +119,11 @@ public class MainController {
 	}
 	
 	@PutMapping("/pizzas/{id}")
-	public String updatePizza(@Valid @ModelAttribute Pizza pizza, BindingResult result, @PathVariable Long id, Model model  ) {
-		
+	public String updatePizza(@Valid @ModelAttribute Pizza pizza, BindingResult result, @PathVariable Long id, Model model, RedirectAttributes redirectAttributes, HttpSession session  ) {
+		if (session.getAttribute("loggedInUser") == null) {
+			redirectAttributes.addFlashAttribute("notLoggedIn", "You must be logged in to view this page");
+			return "redirect:/auth/login";
+		}
 		if (result.hasErrors()) {
 			String[] pizzaSizes = {"Small", "Medium", "Large", "Extra Large"};
 			model.addAttribute("pizzaSizes", pizzaSizes);
@@ -104,7 +135,11 @@ public class MainController {
 	}
 	
 	@DeleteMapping("/pizzas/{id}")
-	public String deletePizza(@PathVariable Long id) {
+	public String deletePizza(@PathVariable Long id, RedirectAttributes redirectAttributes, HttpSession session) {
+		if (session.getAttribute("loggedInUser") == null) {
+			redirectAttributes.addFlashAttribute("notLoggedIn", "You must be logged in to view this page");
+			return "redirect:/auth/login";
+		}
 		pizzaService.deletePizza(id);
 		return "redirect:/";
 	}
